requirements.txt
Django==1.11.10
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.4.0
PyJWT==1.5.3
pytz==2017.3
